feat: add zooming

+++ active_learning.py
@@ -9,7 +9,7 @@ import numpy as np
 class BoundingBox:
     """Class to manage bounding boxes."""
 
-    def __init__(self, x_min: int, y_min: int, x_max: int, y_max: int, class_id: int):
+    def __init__(self, x_min: float, y_min: float, x_max: float, y_max: float, class_id: int):
         self.x_min = x_min
         self.y_min = y_min
         self.x_max = x_max
